Publish v2.4.296

+++ VpnHood.Server.App.Net/_publish.ps1
@@ -1,3 +1,5 @@
+param([switch]$pushDocker)
+
 . "$PSScriptRoot\..\Pub\Common.ps1"
 $packageName = "VpnHoodServer";
 New-Item -Path $packagesServerDir -ItemType Directory -Force
@@ -17,7 +19,7 @@ $linuxScript  | Out-File -FilePath "$packagesServerDir/install-linux.sh" -Encodi
 	-packageDownloadUrl "https://github.com/vpnhood/VpnHood/releases/latest/download/$packageName.zip"
 
 
-# server install-linux2.sh
+# server VpnHoodServer.docker.sh
 $linuxScript = (Get-Content -Path "$PSScriptRoot/Install/VpnHoodServer.docker.sh" -Raw).Replace('$composeUrlParam', "https://github.com/vpnhood/VpnHood/releases/download/$versionTag/VpnHoodServer.docker.yml");
 $linuxScript = $linuxScript -replace "`r`n", "`n";
 $linuxScript  | Out-File -FilePath "$packagesServerDir/VpnHoodServer.docker.sh" -Encoding ASCII -Force -NoNewline;
@@ -25,12 +27,15 @@ $linuxScript  | Out-File -FilePath "$packagesServerDir/VpnHoodServer.docker.sh" 
 # copy compose file
 Copy-Item -path "$projectDir\Install\VpnHoodServer.docker.yml" -Destination "$packagesServerDir\" -Force
 
-# remove old docker containers from local
-$serverDockerImage="vpnhood/vpnhoodserver";
-docker rm -vf $(docker ps -a -q --filter "ancestor=$serverDockerImage")
-docker rmi -f $(docker images -a -q "$serverDockerImage")
+if ($pushDocker)
+{
+	# remove old docker containers from local
+	$serverDockerImage="vpnhood/vpnhoodserver";
+	docker rm -vf $(docker ps -a -q --filter "ancestor=$serverDockerImage")
+	docker rmi -f $(docker images -a -q "$serverDockerImage")
 
-# create name image
-docker build "$solutionDir" -f "$projectDir\Dockerfile" -t ${serverDockerImage}:latest -t ${serverDockerImage}:$versionTag
-docker push ${serverDockerImage}:latest
-docker push ${serverDockerImage}:$versionTag
+	# create name image
+	docker build "$solutionDir" -f "$projectDir\Dockerfile" -t ${serverDockerImage}:latest -t ${serverDockerImage}:$versionTag
+	docker push ${serverDockerImage}:latest
+	docker push ${serverDockerImage}:$versionTag
+}
